CHROMIUM: thermal: mediatek: forward port svs driver for mt8173
This restores the SVS portions of the mtk-thermal driver that was used
in 3.18, and uses them for mt8173. The current state is that thermal and
SVS are split into two separate drivers. However, the current SVS driver
only supports mt8183 [1]. There is no patchset for adding support for
mt8173 to the new driver, therefore this is a pragmatic approach to
restoring SVS to mt8173.

b:157105984 has been filed to track upstream SVS support for mt8173.

[1] https://lore.kernel.org/linux-mediatek/20200316072316.7156-1-roger.lu@mediatek.com/

BUG=b:144478073,b:157105984
TEST=Run power_LoadTest.1hour

Cq-Depend: chromium:2215661
Change-Id: Ie67bb364a96fddb5668d1c2958f648933e9022e1
Signed-off-by: Anand K Mistry <amistry@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2195581
Commit-Queue: Anand Mistry <amistry@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>