CHROMIUM: media: mtk-vcodec: ignore used flag in stateless codec
The used flag is used in the stateful decoder to control whether
the firmware is not using buffers anymore as reference frames.
In a stateless decoder context, reference frames are managed by
userspace, so this flag is useless beyond sanity checks. Since
maintaining its state introduced bugs in both the stateful and
stateless decoders, let's just get rid of it in the stateless
code.

BUG=b:156305939
BUG=chromium:1023703
TEST=VDtests pass on Hana and Kukui.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Change-Id: Iaee5d462224cae14de90a2fa672169c2ee0e4228
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2211704
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>