Revert "UPSTREAM: Input: elants_i2c - report resolution information for touch major"
This reverts commit 03048d5a2f4a732a4939e04f61cb1f22da4ead87.

This commit is causing b/158477364 on Dru and part of b/159195016 on
Waddledoo, so it seems that there are multiple ELAN touchscreens for
which the resolution of 1 unit per mm is incorrect. Reverting on all
Kernel versions seems like a prudent step until we have a more universal
way to set the resolution.

BUG=b:158477364,b:159195016
TEST=none

Signed-off-by: Harry Cutts <hcutts@chromium.org>
Change-Id: Idc26a2b08751092fb42be91cc0104d6b59293bb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2252878
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>