CHROMIUM: iwl7000: mvm: re-enable TX after channel switch
The FW relies on the re-enablement of the TX in order to know
it can exit quiet mode. Currently in case of CSA with quiet mode,
the quiet mode is never cancelled, resulting in a complete traffic
hang.

Change-Id: I09550fdadb61f899242d7e7b7578672372e2b7fe
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: 2f4d947a6c1f448f6dc3c1765aad46069844dbd9