CHROMIUM: Bluetooth: btmrvl: Fix use-after-free when forcing a replug of the SDIO card
Commit 3bfd7663b760 ("CHROMIUM: Bluetooth: btmrvl: Detect hangs and
force a replug of the SDIO card") introduced workqueue task that forces
a replug of the SDIO card when an error is detected. When the card is
removed, it is possible that a reset work task is pending. If that task
runs after the removal has completed, the access of the btmrvl_sdio_card
will result in a use-after-free since the card has been deallocated.

BUG=b:157434249
TEST=On elm-kernelnext, run
     'cts-tradefed run commandAndExit cts -m CtsBluetoothTestCases'
     100 times => system does not kernel panic and reboot

Change-Id: I005aa389cf0505490cad448f99a3a05f9989569b
Signed-off-by: Anand K Mistry <amistry@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2265535
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Anand K Mistry <amistry@chromium.org>
Commit-Queue: Anand K Mistry <amistry@chromium.org>