CHROMIUM: iwl7000: mvm: don't set PSC channels with ssid bitmap
In case of scheduled scan request, PSC channels are added to the
scan request, and are expected to be scanned as part of the
UHB scan after legacy channels scan is completed.
Currently PSC channels flags are set with ssid bitmap value which
is wrong as they should be scanned based on reported collocated APs.
Fix it.

Change-Id: I390e29ecc92fc8266f6486ea5eb428a7cbfea22d
Signed-off-by: Ayala Beker <ayala.beker@intel.com>
iwl7000-tree: 87a862d00f5e4ad85e299da14e3a6f62251f4273