Revert "CHROMIUM: bluetooth: set advertising intervals"
This reverts commit d4e59ea0a52d6670d4c821644f4469869a166692.

Reason for revert - It appears that there was some kind of v4.19 uprev
that caused the Set Advertising Interval feature to be split across a
few commits. We recently reverted most of it, but there were a couple
things that were missed. This CL reverts the remaining artifacts, which
were causing failures in advertising tests on 4.19 when hardware
offloaded multi-advertising was not available.

BUG=b:161830108
TEST=AdvSanity on kukui
Signed-off-by: Daniel Winkler <danielwinkler@google.com>

Change-Id: I4a7b0aab25e38169b27164907672c6bdaf01d905
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2310744
Reviewed-by: Alain Michaud <alainm@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>