CHROMIUM: arm64: dts: mt8183-elm: Set GPU power regulator to always on
Keep the da9211 BUCKB always-on. This works around an issue on Elm/Hana
where sometimes, the regulator is disabled before scpsys is suspended,
causing the suspension of scpsys to fail.

Usually, the GPU and scpsys are suspended by the runtime PM before the
system is suspended, due to the GPU being idle. In this case, scpsys is
suspended inline with the GPU suspend, which then disables the
regulator. However, if the GPU is still active when system is suspended,
GPU suspend occurs but defers suspending scpsys to the PM's noirq phase.
Since GPU suspend disables the regulator, scpsys isn't powered and
suspending it fails with the following error:
[  523.773227] mtk-scpsys 10006000.scpsys: Failed to power off domain mfg_2d

On resume, scpsys is resumed in the noirq phase. Since scpsys requires
power from the regulator, which is still disabled at this point,
attempting to turn it on will hang the CPU when it's performing MMIO. A
HW watchdog eventually reboots the system.

The obvious solution would be to add a link to the regulator from scpsys
in the devicetree. This would prevent the regulator from being disabled
until scpsys is suspended. However, in the case where suspending scpsys
is deferred to the noirq phase, disabling the regulator will fail since
it is connected over I2C which requires IRQs to be enabled. Even in the
usual case where scpsys is suspended inline with the GPU, PM will always
attempt to resume scpsys in noirq. This will attempt to enable the
regulator, which will also fail due to being unable to communicate over
I2C.

Since I2C can't be using with IRQs disabled, a workaround is to never
turn off the regulator.

BUG=b:160941134
TEST='suspend_stress_test -c 100`

Change-Id: Icd29ce3eb667d8c9676431d8b760a2667b52a643
Signed-off-by: Anand K Mistry <amistry@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2317129
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Anand K Mistry <amistry@chromium.org>
Tested-by: Anand K Mistry <amistry@chromium.org>