CHROMIUM: MALI: platform: Fix voltage transitioning logic
There are 2 issues in existing platform code:
- Voltage may be set to invalid low values before frequency is turned
  down.
- Current voltage is not updated properly during voltage transitioning.

And they crash system when we moved to Bifrost r25p0 release.

On r14p0 they doesn't break the system for some reasons, but it's still
better to fix them.

In addition, there's one doubtful code piece but turns out it works
perfectly on MT8183 (unsure if it's intended), so adding explanation for
for ones who will be reading this part as well.

BUG=b:161663868
TEST=log in guest mode on Krane, repeatedly switching between Chrome and
     home page to trigger GPU frequency transitioning from 3e8 to 8e8 and
     reverse
TEST=run webGL aquarium without issue
TEST=manually modify devfreq values to see freq transitioning works, and
     voltage bias is legit all the time

Change-Id: Ia696874d3ea6d4348ca42aa4a90207b7b43fa94b
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2321888
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>