CHROMIUM: MALI: platform: Revise get_step_volt()
The original logic of getting step voltage works exclusively with
current MT8183 OPP table, but that does not guarantee to work with other
OPP tables, in addition it was confusing to always use step_volt[0] in
calculation (although it works).

Revising the part to make it more compatible to handle general cases.

BUG=b:162191647
TEST=log in guest mode on Krane, repeatedly switching between Chrome and
     home page to trigger GPU frequency transitioning from 3e8 to 8e8 and
     reverse
TEST=run webGL aquarium without issue
TEST=manually modify devfreq values to see freq transitioning
     works, and voltage bias is legit all the time

Change-Id: I88de6a396f91a080ca9bd448a504a62b7651b36c
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2321889
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>