CHROMIUM: iwl7000: support firmware reset handshake
There are some races in the hardware that can possibly lead to
a bus lockup later during a restart when we manage to kill the
firmware at a bad time (while it's accessing the bus).

To work around this, add support for a new handshake between
firmware and driver to ensure that the firmware is in a well-
known state before we kill it.

Change-Id: I13de65e0ffcb4186dd4c1a465f66df2e98c9a947
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 0210923ce2e8e3688de04f7656e2732cbb6f35d3