CHROMIUM: img-rogue: 1.10: re-add mtk_mfg_disable_hw_apm()
Clear the apm hw register when there is a power down transition.
Before this change, MTK scpsys and the GPU firmware could attempt
to drive the power to the MFG power island and expose the system
to a potential race. Sometime manifesting with the following
message:

  [    1.870134] mtk-scpsys 10006000.scpsys: Failed to power off domain mfg

This is due to a timeout in the scpsys sram disable path, the
poll succeeds if manually clearing APM registers in
mtk_mfg_disable(), and partially reverting 6ed1b38d10df.

BUG=b:168839892
TEST=Boot to UI on Elm

Change-Id: I13108ff8a5376bd9f546e3f36dd57f546e898872
Signed-off-by: Luigi Santivetti <luigi.santivetti@imagination.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2503691
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>