FROMLIST: media: platform: Add mechanism to handle jpeg hardware's locking up
There is a delayed work scheduled before starting the hardware and
canceled in the interrupt handler. If the delayed work is executed, it
resets the hardware and reports the failure to V4L2, so that the
execution can continue from next frames.

Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Xia Jiang <xia.jiang@mediatek.com>

(am from https://patchwork.kernel.org/patch/11713725/ )
BUG=b:134367768
TEST=build and image ok.

Signed-off-by: Maoguang Meng <maoguang.meng@mediatek.com>
Change-Id: I73ec0583a8defce0fefd5058e39b9df2a6b4ab8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2362207
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Maoguang Meng <maoguang.meng@mediatek.corp-partner.google.com>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Maoguang Meng <maoguang.meng@mediatek.corp-partner.google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>