UPSTREAM: media: mtk-vcodec: venc: use platform data for ENUM_FRAMESIZES
vidioc_enum_framesizes() assumes that all encoders support H.264 and VP8,
which is not necessarily true and requires to duplicate information about
the supported codecs which is already stored in the platform data.

Fix this by referring to the platform data to find out whether a given
format is supported. Since the supported sizes are all the same
regardless of the format, we can then return a copy of a static value if
the format is supported.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Acked-by: Tiffany Lin <tiffany.lin@mediatek.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit cd385511520e22ec0e8bdbfb7fbb997a4256ac88)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

BUG=b:149437087
TEST=tast video.EncodeAccelNew.h264_* passes on Krane
TEST=tast video.EncodeAccelNew.h264_* passes on Hana

Change-Id: Ic247afe47bf254d1aed07df30fa274bc87435749
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2667299
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>