CHROMIUM: iwl7000: mac80211: add fragment cache to sta_info
Prior patches protected against fragmentation cache attacks
by coloring keys, but this shows that it can lead to issues
when multiple stations use the same sequence number. Add a
fragment cache to struct sta_info (in addition to the one in
the interface) to separate fragments for different stations
properly.

This then automatically clear most of the fragment cache when a
station disconnects (or reassociates) from an AP, or when client
interfaces disconnect from the network, etc.

On the way, also fix the comment there since this brings us in line
with the recommendation in 802.11-2016 ("An AP should support ...").
Additionally, remove a useless condition (since there's no problem
purging an already empty list).

BUG=b:187853103, b:187853926
TEST=Verified wifi_matfunc/perf.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 5894b7243671f337bbdd1fc7ac55ca3a980eaa48
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I1ef101e3784d13e8f6600d83de7ec9a3a45bcd52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2888814
Reviewed-by: Brian Norris <briannorris@chromium.org>