CHROMIUM: HID: microsoft: Add Xbox One S controller to microsoft hid driver
Add the Xbox One S controller to be handled by the microsoft hid driver.
In later kernel versions this controller is added to the driver to add
support for force feedback. See
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/73c5b254c36529c84c9d19e07905f7103bb32e79

Without this change the Xbox Home button is put in its own input and it
is not functional.

This change has the effect of not setting the HID_QUIRK_INPUT_PER_APP
bit on this device which will correctly add the Xbox Home button to
input with the rest of the buttons.

BUG=b:168800073
TEST=Connect Xbox One S controller over bluetooth
     Observe that the Xbox Home button works as expected on "Game Pass"
     app

Change-Id: I49f96a7b28a8a8b544740f9fdc5a185e6a95ae87
Signed-off-by: Mathew King <mathewk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2425159
Reviewed-by: Harry Cutts <hcutts@chromium.org>