BACKPORT: FROMGIT: Bluetooth: Resume advertising after LE connection
When an LE connection request is made, advertising is disabled and never
resumed. When a client has an active advertisement, this is disruptive.
This change adds resume logic for client-configured (non-directed)
advertisements after the connection attempt.

The patch was tested by registering an advertisement, initiating an LE
connection from a remote peer, and verifying that the advertisement is
re-activated after the connection is established. This is performed on
Hatch and Kukui Chromebooks.

Signed-off-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

(cherry picked from commit 2943d8ede38310db932eb38f91aa1094b471058c
 git://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git/ master)

Conflicts:
   net/bluetooth/hci_request.h

Conflict caused by address resolution declaration that exists upstream
but not in Cros.

BUG=b:162324887
TEST=All LE and SR autotests on hatch and kukui

Change-Id: I8362b4cedb0f34b7a88b8dbd3a62155085e02ea7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2521851
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>