CHROMIUM: iwl7000: mvm: don't crash if registration failed
If we failed to register to mac80211, we must not unregister
the vendor commands. This has already been done in the failure
path of the registration to mac80211. Check first that the
registration to mac80211 went well and avoid to crash if it
didn't.

Change-Id: I3cec41e35ee5bda2207ff6994044440565951390
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 257f9b0c03867f1a606c4c20a7bb14ba477cca50