BACKPORT: pinctrl: mtk: Fix low level output voltage issue
This patch is used to fix low level output voltage issue.
A pin is changed from input pull-up to output high.
The Dout value of the pin is default as 0.
If we change the direction of the pin before the dout value of the pin,
It maybe produce a low level output voltage between "input pull-up" and
"output high".

Signed-off-by: Zhiyong Tao <zhiyong.tao@mediatek.com>
Link: https://lore.kernel.org/r/20201120093058.7248-2-zhiyong.tao@mediatek.com
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

(cherry picked from commit 517c3f5a8683c950efe17aa01e55efb3b0f2c770)
[drinkcat: conflicts in drivers/pinctrl/mediatek/pinctrl-paris.c,
the additional if (err) is not actually needed and is cleaned up
upstream]

BUG=b:159688118
TEST=boot Kukui

Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Change-Id: Ied5aed0ad7c759adef54e916f219dd2d3518d047
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2573184
Tested-by: Ikjoon Jang <ikjn@chromium.org>
Reviewed-by: Ikjoon Jang <ikjn@chromium.org>
Commit-Queue: Ikjoon Jang <ikjn@chromium.org>