CHROMIUM: iwl7000: mac80211: assure all fragments are encrypted
Do not mix plaintext and encrypted fragments in protected Wi-Fi
networks. This fixes CVE-2020-26147.

Previously, an attacker was able to first forward a legitimate encrypted
fragment towards a victim, followed by a plaintext fragment. The
encrypted and plaintext fragment would then be reassembled. For further
details see Section 6.3 and Appendix D in the paper "Fragment and Forge:
Breaking Wi-Fi Through Frame Aggregation and Fragmentation".

Because of this change there are now two equivalent conditions in the
code to determine if a received fragment requires sequential PNs, so we
also move this test to a separate function to make the code easier to
maintain.

BUG=b:187853103, b:187853926
TEST=Verifies wifi_matfunc/perf.

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: d587501457ae6a44a48fa4779fb17f11f0a8b0c0
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I5acfdb552cc1d20c339c262315950b3eac491397
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2888810
Reviewed-by: Brian Norris <briannorris@chromium.org>