CHROMIUM: iwl7000: pcie: invert values of NO_160 device config entries
The NO_160 flag specifies if the device doesn't have 160 MHz support,
but we errorneously assumed the opposite.  If the flag was set, we
were considering that 160 MHz was supported, but it's actually the
opposite.  Fix it by inverting the bits, i.e. NO_160 is 0x1 and 160
is 0x0.

Change-Id: I83884286b688965293e9810381808039bd7eedae
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 74d0e8bb6d775c5db9f3eaad11e6693e85dfb7c2