CHROMIUM: iwl7000: pnvm: increment the pointer before checking the TLV
If the SKU_ID doesn't match, we don't increment the pointer and keep
checking the same TLV over and over again.

We need to increment the pointer in all situtations, namely if the TLV
is not a SKU_ID, if the SKU_ID matched or if the SKU_ID didn't match.
So we can increment the pointer already before checking for these
conditions to solve the problem.

Change-Id: I9a5a54906cf0f4ec8af981d6066bfd771152ffb9
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: f3411d2333300b9df5b727e535b8e622f6d159a7