CHROMIUM: soc: mediatek: Drop parent/child relationship for mfg domains on mt8173
Based on commit 4762f11e0c95 ("CHROMIUM: soc: mediatek: SCPSYS: Drop
parent/child relationship for mfg domains")

This change drops the genpd relationships between the various mfg power
domains and instead treats them as a single domains to be powered off/on
together. This also ignores errors when failing to power off any of the
mfg domains.

This is workaround for system freeze/hang that occurs when the mfg_2d
domains fails to power off, and should be reverted once the root cause
is found. This is the behaviour of 3.18 which experiences mfg_2d power
off failures, but the devices appear to continue functioning.

BUG=b:168839892
TEST=Build/deploy to hana. Run webgl aquarium, watch some youtube.

Change-Id: Ibd47bedb1b19afee319e854152e066f43212f8f0
Signed-off-by: Anand K Mistry <amistry@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2589179
Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>