CHROMIUM: Bluetooth: Ignore global Connectable setting for extended adv
Existing logic enforces a "connectable" advertisement if the global
Connectable setting is active, even if the advertisement requested was
non-connectable. This can cause a case where the client requires a
random address, but kernel will ignore that request. This patch removes
the check for the global setting, to ensure we provide a random address
if the client did not specify Connectability in their advertisement
request.

Change was tested by verifying that Nearby startup flow uses a random
address in its advertisement as expected.

BUG=b:175069101
TEST=Test Nearby startup flow on hatch

Signed-off-by: Daniel Winkler <danielwinkler@google.com>
Change-Id: I99ea32b02b50ec996239e1a4e489cba47ae90a62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2593511
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>