CHROMIUM: iwl7000: mvm: don't WARN if we can't remove a time event
It is not very useful to WARN if we can't send a host command
The firmware is likely in a bad situation and the fact that
we didn't send the host command has an impact on the firmware
only, not on the driver. The driver could clean up all its
state.

Don't WARN in this case, but just leave a smaller note.

Change-Id: I4f9b769bc38d68f8ed43f77d2cd75e8f1993e964
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
(cherry picked from commit c1fdc7cc74904ca54ea47a71561639921b1e4f58)
iwl7000-tree: 4ba548116134f898e65141a9c0198c7cb14f44ee