CHROMIUM: iwl7000: fw: fix powersave debugfs out-of-bounds read
The code here checked only that the beginning of the field
(offsetof) fits into the data returned by the firmware, not
that the *entire* field fits. Fix this to use offsetofend()
and see that is <= the size to fix this.

Change-Id: If411638e985242036476192620442cbd179d9a30
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: db4ab03add49350eb6df7813e790be95f1a848dc