CHROMIUM: media: mtk-vcodec: venc: align with upstream code
This is a non-functional change that just aligns the encoder code with
the version that has landed upstream. Since the encoder landed in
Chromium OS first, the upstream version differs a bit.

The changes are indentation adjustments, renaming, and a bit of code
flow that does not change the logic.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

BUG=b:149437087
TEST=tast video.EncodeAccelNew.h264_* passes on Krane
TEST=tast video.EncodeAccelNew.h264_* passes on Hana

Change-Id: I1953458b55666dc8df1d079f288cbf03d103a127
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2668107
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>