CHROMIUM: media: align H.264 stateless interface with upstream
For a long time Chrome OS has used its own H.264 stateless controls,
mostly because the upstream versions were non-existing or not stabilized
yet. As of Linux 5.11 and commit 46a309d27517 ("media: uapi: move H264
stateless controls out of staging"), the controls are officially stable.
This CL aligns our codebase with the upstream definitions.

This CL is a bit all over the place, but that's unfortunately inevitable
since the controls affect the V4L2 core as well as the hantro and
mtk-vcodec drivers. The mentions of "upstream" below refer to commit
8f202f8e9ff3.

The following is a summary of the changes:

* Obsolete controls documentation is removed from ext-ctrls-codec.rst.
  New controls documentation is not added to limit the amount of
  changes. Please refer to upstream for documentation.
* h264-ctrls.h now contains a verbatim copy of the upstream H.264
  controls.
* v4l2-ctrls.c is updated to the upstream code related to
  these controls.
* v4l2-h264.[ch], which compute the reference lists, are now perfectly
  aligned with upstream.
* The mtk-vcodec driver is updated to use this new ABI and confirmed to
  work,
* The hantro driver is updated to use this new ABI. Thankfully in its
  case the only requirement is that control names and structure members
  are properly updated.

BUG=b:149437087
TEST=Video decoding works on Kukui
TEST=Video decoding works on Hana

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Change-Id: Iebd495db58859aafabc147342ac6aaf9dbabeae8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2719247
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>