CHROMIUM: MALI: Handle context termination during system suspend
Patch from ARM ticket: 00244659 (GPUCORE-24808)

It was earlier assumed that context termination cannot kick in
after the system suspend has started and userspace threads have
been frozen.
But it has been seen on Customer side, that context termination
can happen (due to the killing of an App) whilst the system suspend
of the GPU device is ongoing or has completed. And a hang can occur
if the context termination is not blocked until the resume of GPU device.

This commit updates the kbase_destroy_context() function for JM
GPU to wait for the resume of GPU device.

The actual reason for the hang is not known yet, as the similar
issue couldn't be reproduced locally.

TI2: 645240 (PLAN-24380r2 R28 DDK Precommit)

BUG=b:176946362
TEST=emerge-kukui sys-kernel/chromeos-kernel-4_19

Signed-off-by: Nick Fan <Nick.Fan@mediatek.com>
Change-Id: I3e1ba0c98d1d1d7bbc363d899c79c28f55c48ba3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2719780
Tested-by: Fei Shao <fshao@chromium.org>
Auto-Submit: Fei Shao <fshao@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>