CHROMIUM: Input: elants_i2c: Default to low level interrupt for Chromebooks
If device uses edge triggered interrupts, it may happen that it sends
it before its IRQ is enabled (enable_irq()). Kernel normally should
replay such intterupts, however, if the relevant IRQ chip does not
implement irq_retrigger() callback, and software resending is not used
(HARDIRQS_SW_RESEND is not selected), it does not happen.
If the device keeps the interrupt line low, no new interrupt gets
detected, and hence the device is unreponsive, as we do not issue any
IRQ handler, since we do not notice new IRQs.

This is what indeed happens on Relm chromebook considering the
touchscreen. If the laptop gets suspended, the touchscreen is not
working after resuming as the touchscreen lowers the intterupt line
before we re-enable the interrupt.

This issue can be worked-around if we use level interrupt.

BUG=b:162276454
TEST=Compile, upload the kernel to Relm chrombeook. Verify that the
touchscreen works after resuming.

Signed-off-by: Angela Czubak <acz@semihalf.com>
Change-Id: I5f63937dca04570d9801126c07c1c766d8a856a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2720240
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>
Tested-by: Angela Czubak <aczubak@google.com>