HACK: CHROMIUM: media: mtk-mdp: Compute plane address with data_offset
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.

BUG=b:172216855
BUG=b:172223525
TEST=video_encode_accelerator_tests on elm

Change-Id: I254c39d0355f2f7122dc4d0923f6bbcfbe7de63e
Signed-off-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2726264
Reviewed-by: Tomasz Figa <tfiga@chromium.org>