CHROMIUM: iwl7000: add a framework to mock up ACPI method calls
Create a framework, enabled during compilation time, to replace
actuall calls to ACPI methods with mock ups that return values defined
in the driver.

This is useful in order to test ACPI features more easily, without
having to use modified BIOSes with the ACPI tables we want to test.

For now, we have a single hard-coded table that can be used to
validate the mock-up functionality, but this will be replaced in later
patches with dynamic lists generated from binary files.

Change-Id: I620faa7031133e328e335f9e527a1b55b968c79f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 1d64252aecc4c1f4139302a2a2f7c81612cd279f