CHROMIUM: Input: atmel_mxt_ts - print message when device not present
Based on some user logs it looks as if checking the presence of the
touchpad or touchscreen might fail at times.
Add an explicit message so that we do not guess this happens based on
the fact that the interrupt has not been set up or there are no more
logs form the driver.

BUG=b:183914549
TEST=Compile, upload the kernel to Caroline chromebook, verify it boots
fine and the touchpad and touchscreen work.

Signed-off-by: Angela Czubak <acz@semihalf.com>
Change-Id: Ib3d2b594fe117170d4b241fd1b2d39e91a8feaee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2810609
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>