CHROMIUM: MALI: Set short timeout for kbase DMA fence
It's still unclear why some Mali kbase DMA fences never get signaled,
but when that happens, the screen freezes and the system reboot in the
end.
This is just a temporary hack to use a shorter timeout for the wait
function, so the driver can at least fall back to some cleanup logic
and unblock the frozen screen.

BUG=b:175656896
TEST=system does not reboot after screen hang when running Android
     apps or CTS tests

Change-Id: Ia9bd13be2f3739d913c5b1602e88c09aca569390
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2823449
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>