CHROMIUM: Notify L2CAP when hci_chan is freed
Consult the bug. It is suspected that l2cap_chan still access
hci_chan, but hci_chan is freed, so we crashed. This CL try to mark
hci_chan as NULL when it is freed.

Since I cannot reproduce the crash easily, it's hard to confirm the
correctness of this CL. Therefore, let's just make it a CHROMIUM
only patch and see whether this could help decreasing the number
of crashes. Therefore, it's also not structured like one is expecting
for an upstreamable CL.

BUG=b:168048314
TEST=run autotest

Signed-off-by: Archie Pusaka <apusaka@chromium.org>

Change-Id: I4d9900cf90b75d88fd49cfde8d6c2fca407917aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2855885
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>